bump version, add remote file exists check before running check

+++ lib/capistrano/unicorn.rb
@@ -31,19 +31,21 @@ module Capistrano
         end
         desc "graceful stop unicorn"
         task :graceful_stop, :roles => :app, :except => { :no_release => true } do
-          if pid_running?(unicorn_pid)
+          if remote_file_exists(unicorn_pid) && pid_running?(unicorn_pid)
             run "kill -s QUIT `cat #{unicorn_pid}`"
           end
         end
         desc "graceful stop OLD unicorn"
         task :graceful_stop_old, :roles => :app, :except => { :no_release => true } do
-          if pid_running?(unicorn_old_pid)
+          if remote_file_exists(unicorn_old_pid) && pid_running?(unicorn_old_pid)
             run "kill -s QUIT `cat #{unicorn_old_pid}`"
           end
         end
         desc "reload unicorn"
         task :reload, :roles => :app, :except => { :no_release => true } do
-          if pid_running?(unicorn_pid)
+          
+          if remote_file_exists?(unicorn_pid) && pid_running?(unicorn_pid)            
+          # if pid_running?(unicorn_pid)
             puts "UNICORN RUNNING, reloading"
             run "kill -s USR2 `cat #{unicorn_pid}`"
             unicorn.graceful_stop_old
